placed Reid's and my code together
after I got the movement done 100% I merged the coding together instead of doing his pull request to the old code. I looked at what he changed and placed it in there for us to use. movement is all done and the level design is what i'm currently working on.