Changed layout
I've changed the layout of the game to make it easier to add sprites in, aswell I have also added mobs into the game that can kill you, they only move side to side which is exactly what I wanted. this level is close to done, but with the new design I can now change the levers to make them do something when you hit them so you know that thy have already been hit.